# swagger.yaml
...
swagger: "2.0"
host: "localhost:8000"
basePath: "/api/v0.3"
info:
  title: User-Organization API
  description: An API which supports management of customers and organizations within the LEXIS portal; this version is intended for use in LEXIS Portal R1
  version: "0.3.2021-12-15"
  contact:
    email: diego@cyclops-labs.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
securityDefinitions:
  APIKeyHeader:
    type: apiKey
    in: header
    name: X-API-KEY
  APIKeyParam:
    type: apiKey
    in: query
    name: api_key
  Keycloak:
    type: oauth2
    flow:  accessCode
    authorizationUrl: 'http://localhost:8080/auth/realms/Dev/protocol/openid-connect/auth'
    tokenUrl: 'http://localhost:8080/auth/realms/Dev/protocol/openid-connect/token'
    scopes:
      lex_adm: $Deity scope, can do absoulutely anything
      lex_sup: $Deity scope with restrictions.
      org_mgr: Organization Manager
      prj_mgr: Project/Workflows Manager
      dat_mgr: Datasets Manager
      end_usr: Plain user
schemes:
  - http
  - https
security:
  - Keycloak: [lex_adm, lex_sup, org_mgr, prj_mgr, dat_mgr, end_usr]
paths:
...
  /hpc/resource:
    get:
      tags:
        - hpcManagement
      produces:
        - application/json
      summary: list HPCResource
      operationId: listHPCResources
      description: lists all HPCResource
      responses:
        '200':
          description: list of HPCResources
          schema:
            type: array
            items:
              $ref: "#/definitions/HPCResource"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: query
          name: scope
          description: organization scope switch
          type: string
          enum:
          - ALL
          - OWN
          - ALLOWED
          default: ALL
    post:
      tags:
        - hpcManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: create HPCResource
      operationId: createHPCResource
      description: Creates a new HPCResource
      responses:
        '201':
          description: HPCResource created
          schema:
            $ref: "#/definitions/CreatedResponse"
        '409':
          description: an existing item already exists
          schema:
            $ref: "#/definitions/ConflictResponse"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: body
          name: HPCResource
          description: HPCResource to be added
          schema:
            $ref: '#/definitions/HPCResource'
  /hpc/resource/{id}:
    get:
      tags:
        - hpcManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Get specific HPCResource
      operationId: getHPCResource
      description: get HPCResource with given id
      responses:
        '200':
          description: HPCResource returned
          schema:
            $ref: "#/definitions/HPCResource"
        '404':
          description: HPCResource with not found
          schema:
            $ref: "#/definitions/MissingResponse"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: Id of HPCResource to be obtained
    put:
      tags:
        - hpcManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Update specific HPCResource
      operationId: updateHPCResource
      description: Update HPCResource with given id
      responses:
        '200':
          description: updated HPCResource
          schema:
            $ref: "#/definitions/HPCResource"
        '404':
          description: HPCResource with not found
          schema:
            $ref: "#/definitions/MissingResponse"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: Id of HPCResource to be obtained
        - in: body
          name: HPCResource
          required: true
          description: updated HPCResource data to be added
          schema:
            $ref: "#/definitions/HPCResource"
    delete:
      tags:
        - hpcManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Delete specific HPCResource
      operationId: deleteHPCResource
      description: Delete HPCResource with given id
      responses:
        '200':
          description: deleted HPCResource
          schema:
            $ref: "#/definitions/DeletedResponse"
        '404':
          description: HPCResource with not found
          schema:
            $ref: "#/definitions/MissingResponse"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: Id of HPCResource to be obtained
...
definitions:
...
  HPCResource:
    type: object
    properties:
      HPCResourceID:
        type: string
        x-go-custom-tag: gorm:"column:hpcresourceid;primary_key;unique;default:md5(random()::text || clock_timestamp()::text)::uuid"
      AssociatedHPCProject:
        type: string
        x-go-custom-tag: gorm:"column:associatedhpcproject"
      AssociatedLEXISProject:
        type: string
        format: uuid
        x-go-custom-tag: gorm:"column:associatedlexisproject;type:uuid"
      ApprovalStatus:
        type: string
        enum:
        - ACCEPTED
        - REJECTED
        - PENDING
        x-go-custom-tag: gorm:"column:approvalstatus"
      CloudNetworkName:
        type: string
        x-go-custom-tag: gorm:"column:cloudnetworkname"
      ProjectNetworkName:
        type: string
        x-go-custom-tag: gorm:"column:projectnetworkname"
      HEAppEEndpoint:
        type: string
        x-go-custom-tag: gorm:"column:heappeendpoint"
      HPCProvider:
        type: string
        enum:
        - IT4I
        - LRZ
        - ICHEC
        x-go-custom-tag: gorm:"column:hpcprovider"
      OpenStackEndpoint:
        type: string
        x-go-custom-tag: gorm:"column:openstackendpoint"
      OpenStackProjectID:
        type: string
        x-go-custom-tag: gorm:"column:openstackprojectid"
      ResourceType:
        type: string
        enum:
        - CLOUD
        - HPC
        - SMARTGW
        x-go-custom-tag: gorm:"column:resourcetype"
      TermsConsent:
        type: boolean
        x-go-custom-tag: gorm:"column:termsconsent;type:bool"
  CreatedResponse:
    type: object
    properties:
      ID:
        type: string
      Link:
        type: string
  ConflictResponse:
    type: object
    properties:
      ID:
        type: string
      Message:
        type: string
  DeletedResponse:
    type: object
    properties:
      ID:
        type: string
      Message:
        type: string
...
  ErrorResponse:
    type: object
    properties:
      Message:
        type: string
  MissingResponse:
    type: object
    properties:
      ID:
        type: string
      Message:
        type: string
...